+++ src/bfabric/experimental/workunit_definition.py
@@ -48,6 +48,7 @@ class WorkunitRegistrationDefinition(BaseModel):
 
     model_config = ConfigDict(extra="forbid")
 
+    application_id: int
     workunit_id: int
     container_id: int
     container_type: Literal["project", "order"]
@@ -56,6 +57,7 @@ class WorkunitRegistrationDefinition(BaseModel):
     def from_workunit(cls, workunit: Workunit) -> WorkunitRegistrationDefinition:
         """Loads the workunit registration definition from the provided B-Fabric workunit."""
         data = {
+            "application_id": workunit["application"]["id"],
             "workunit_id": workunit.id,
             "container_id": workunit.container.id,
             "container_type": workunit.container.ENDPOINT,