+++ app_runner/docs/changelog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Register single file command: `bfabric-app-runner outputs register-single-file`
+- Implement copy resource `UpdateExisting.IF_EXISTS` and `UpdateExisting.REQUIRED` support.
 - `application_id` and `application_name` now part of `WorkunitRegistrationDefinition`.
 
 ## \[0.0.7\] - 2024-11-22