+++ app_runner/docs/changelog.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## \[Unreleased\]
 
+## \[0.0.8\] - 2025-01-08
+
 ### Added
 
 - Register single file command: `bfabric-app-runner outputs register-single-file`